Add base ODE solver class
